fixed add, delete, total for income and expense, added update for tasks

+++ pages/api/income/delete.ts
@@ -2,26 +2,44 @@ import { NextApiRequest, NextApiResponse } from "next";
 import clientPromise from "@/lib/db";
 import { getToken } from "next-auth/jwt";
 import { ObjectId } from "mongodb";
-import { Token } from "@/interfaces";
+
+const secret = process.env.JWT_SECRET;
 
 const deleteIncome = async (req: NextApiRequest, res: NextApiResponse) => {
   if (req.method === "DELETE") {
-    const token = (await getToken({ req })) as Token;
+    const token = await getToken({ req, secret });
+
     if (!token) {
       return res.status(401).end();
     }
 
+    const userId = new ObjectId(token);
+
     const { id } = req.body;
+    if (!id) {
+      return res.status(400).json({ error: "Missing expense ID" });
+    }
+
     const client = await clientPromise;
     const db = client.db("budget-v2");
 
     try {
       const result = await db.collection("income").deleteOne({
         _id: new ObjectId(id),
-        userId: token.sub, // Используем token.sub
+        userId: new ObjectId(userId),
       });
-      res.status(200).json(result);
+      console.log("Delete result:", result);
+
+      if (result.deletedCount === 0) {
+        return res
+          .status(404)
+          .json({ error: "Expense not found or not authorized" });
+      }
+
+      res.status(200).json({ result, message: "Expense deleted successfully" });
     } catch (error) {
+      console.error("Failed to delete expense:", error);
+
       res.status(500).json({ error: "Failed to delete income" });
     }
   } else {
